Keep scroll by keyboard after child focused node is removed
So far, scroll stops if focused child node is removed
while scrolling parent node.
This is because the event handler passes the |MousePressNode|.
If focused child node(|MousePressNode|) is removed,
it will be null and scrollable area changed to document.

|LogicalScroll| finds scrollable area from |MousePressNode|.
If we save this scrollable area node,
we can get one more chance to find proper node to scroll.

Bug: 493078
Change-Id: I6e81ca0a0d15e0def66c8457840ca9e1343e8dba