part11 : modify wpt 'track-element-src-change.html' and 'track-element-src-change-error.html'.
According to the spec [1], the text track list of cues is initially empty. It is dynamically modified when the referenced file is parsed.

Therefore, after reset the track element's url, the number of cues of the text track should be zero, until we start parsing resource and add cue into the list.

[1]
https://html.spec.whatwg.org/multipage/media.html#sourcing-out-of-band-text-tracks:text-track-list-of-cues
https://html.spec.whatwg.org/multipage/media.html#sourcing-out-of-band-text-tracks:text-track-list-of-cues-3

Differential Revision: https://phabricator.services.mozilla.com/D31560

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1550633
gecko-commit: d85485f76233138488e76d8434be46d954547ea3
gecko-integration-branch: autoland
gecko-reviewers: jya