part14 : add a test for 'honor user preferences for automatic text track selection'.
This test is used to ensure that we queue 'honor user preferences for automatic text track selection' as a marco task, not a mirco task.

In this test, we would trigger a media event before queuing a text track selection task, and check the text track's mode to know whether the text track selection runs after the task for media event.

Differential Revision: https://phabricator.services.mozilla.com/D31921

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1550633
gecko-commit: 583d4e0945ad964ef5a9cc1a8ccec8faa248e32b
gecko-integration-branch: autoland
gecko-reviewers: bzbarsky